pyproject.toml
[project]
name = "retrieval-agents"
version = "0.0.1"
description = "RAG agents"
authors = [
    { name = "Cosmac Inc." },
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "langgraph>=0.2.6",
    "langchain-openai>=0.1.22",
    "langchain-ollama>=0.3.1",
    "langchain-anthropic>=0.3.10",
    "langchain-nomic>=0.1.4",
    "langchain-unstructured>=0.1.5",
    "unstructured[md]>=0.17.2",
    "nomic[local]>=3.4.1",
    "langchain>=0.3.22",
    "langchain-fireworks>=0.1.7",
    "python-dotenv>=1.0.1",
    "langchain-elasticsearch>=0.3.2",
    "langchain-pinecone>=0.1.3,<0.3.0",
    "msgspec>=0.18.6",
    "langchain-mongodb>=0.1.9",
    "langchain-cohere>=0.4.3",
    "beautifulsoup4>=4.13.3",
    "langchain-chroma>=0.2.2"
]

[project.optional-dependencies]
dev = [
    "mypy>=1.11.1",
    "ruff>=0.6.1",
    "langgraph-cli[inmem]>=0.1.81",
    "ipykernel>=6.29.5",
    "pytest-cov>=6.1.1",
    "pre-commit>=4.2.0",
    "codespell>=2.4.1"
]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["retrieval_agents.agents", "retrieval_agents"]
[tool.setuptools.package-dir]
"retrieval_agents.agents" = "src/retrieval_agents/agents"
"retrieval_agents" = "src/retrieval_agents"


[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # First line should be in imperative mood
    "T201",
    "UP",
]
lint.ignore = [
    "UP006",
    "UP007",
    # We actually do want to import from typing_extensions
    "UP035",
    # Relax the convention by _not_ requiring documentation for every function parameter.
    "D417",
    "E501",
]
[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]
[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.codespell]
builtin = "clear,rare,code,names"
skip    = ".mypy_cache,.pytest_cache,.ruff_cache,htmlcov,retrieval_agents.egg-info,.coverage,.env,*.ipynb,*.log"
ignore-words = ".codespellignore"